Added deviations estimate to the MLE fitted parameters. This is derived from the Fisher Information matrix and assumes the fit has achieved the Cramer-Roa lower bounds.
Note that the deviations from the least squares estimators are lower
than the deviations from the Fisher matrix. However the Apache LVM
fitter and the custom NonLinearFit class produce the same output so this
is presumed correct, it is just that the fitters have a smaller
confidence interval on the fitted parameters than the Cramer-Roa lower
bounds.

+++ src/main/java/gdsc/smlm/fitting/nonlinear/BaseFunctionSolver.java
@@ -68,7 +68,7 @@ public abstract class BaseFunctionSolver implements FunctionSolver
 	{
 		int[] indices = f.gradientIndices();
 		for (int i = 0; i < indices.length; i++)
-			deviations[indices[i]] = (double) Math.sqrt(covar[i][i]);
+			deviations[indices[i]] = Math.sqrt(covar[i][i]);
 	}
 
 	public static double getSumOfSquares(final int n, double[] y)
